+++ test/analyzers/lint_analyzer/rules/rules_list/avoid_returning_widgets/avoid_returning_widgets_test.dart
@@ -26,22 +26,21 @@ void main() {
 
       RuleTestHelper.verifyIssues(
         issues: issues,
-        startOffsets: [88, 175, 289, 358, 485, 614, 749, 1012],
-        startLines: [6, 11, 20, 25, 30, 35, 41, 55],
-        startColumns: [3, 3, 3, 3, 3, 3, 1, 1],
-        endOffsets: [127, 231, 344, 414, 542, 677, 784, 1087],
+        startOffsets: [755, 791, 827, 859, 892, 961, 292, 1016, 1271],
+        startLines: [36, 38, 40, 42, 44, 47, 16, 53, 67],
+        startColumns: [5, 5, 5, 5, 5, 14, 3, 1, 1],
+        endOffsets: [776, 812, 844, 877, 915, 984, 331, 1071, 1346],
         locationTexts: [
+          '_localBuildMyWidget()',
+          '_getWidgetsIterable()',
+          '_getWidgetsList()',
+          '_getWidgetFuture()',
+          '_buildMyWidget(context)',
+          '_buildMyWidget(context)',
           'Widget get widgetGetter => Container();',
-          'Widget _getMyShinyWidget() {\n'
-              '    return Container();\n'
-              '  }',
-          'Container _getContainer() {\n'
-              '    return Container();\n'
-              '  }',
-          'Iterable<Widget> _getWidgetsIterable() => [Container()];',
-          'List<Widget> _getWidgetsList() => [Container()].toList();',
-          'Future<Widget> _getWidgetFuture() => Future.value(Container());',
-          'Widget _getWidget() => Container();',
+          'Widget _globalBuildMyWidget() {\n'
+              '  return Container();\n'
+              '}',
           '@ignoredAnnotation\n'
               'Widget _getWidgetWithIgnoredAnnotation() => Container();',
         ],
@@ -52,8 +51,9 @@ void main() {
           'Avoid returning widgets from a function.',
           'Avoid returning widgets from a function.',
           'Avoid returning widgets from a function.',
-          'Avoid returning widgets from a function.',
-          'Avoid returning widgets from a function.',
+          'Avoid returning widgets from a getter.',
+          'Avoid returning widgets from a global function.',
+          'Avoid returning widgets from a global function.',
         ],
       );
     });
@@ -63,7 +63,7 @@ void main() {
       final config = {
         'ignored-names': [
           '_getWidgetFuture',
-          '_getWidget',
+          '_buildMyWidget',
         ],
         'ignored-annotations': [
           'ignoredAnnotation',
@@ -74,20 +74,18 @@ void main() {
 
       RuleTestHelper.verifyIssues(
         issues: issues,
-        startOffsets: [88, 175, 289, 358, 485, 814, 879, 944],
-        startLines: [6, 11, 20, 25, 30, 45, 48, 51],
-        startColumns: [3, 3, 3, 3, 3, 1, 1, 1],
-        endOffsets: [127, 231, 344, 414, 542, 877, 942, 1002],
+        startOffsets: [755, 791, 827, 292, 1016, 1073, 1138, 1203],
+        startLines: [36, 38, 40, 16, 53, 57, 60, 63],
+        startColumns: [5, 5, 5, 3, 1, 1, 1, 1],
+        endOffsets: [776, 812, 844, 331, 1071, 1136, 1201, 1261],
         locationTexts: [
+          '_localBuildMyWidget()',
+          '_getWidgetsIterable()',
+          '_getWidgetsList()',
           'Widget get widgetGetter => Container();',
-          'Widget _getMyShinyWidget() {\n'
-              '    return Container();\n'
-              '  }',
-          'Container _getContainer() {\n'
-              '    return Container();\n'
-              '  }',
-          'Iterable<Widget> _getWidgetsIterable() => [Container()];',
-          'List<Widget> _getWidgetsList() => [Container()].toList();',
+          'Widget _globalBuildMyWidget() {\n'
+              '  return Container();\n'
+              '}',
           '@FunctionalWidget\n'
               'Widget _getFunctionalWidget() => Container();',
           '@swidget\n'
@@ -99,11 +97,11 @@ void main() {
           'Avoid returning widgets from a function.',
           'Avoid returning widgets from a function.',
           'Avoid returning widgets from a function.',
-          'Avoid returning widgets from a function.',
-          'Avoid returning widgets from a function.',
-          'Avoid returning widgets from a function.',
-          'Avoid returning widgets from a function.',
-          'Avoid returning widgets from a function.',
+          'Avoid returning widgets from a getter.',
+          'Avoid returning widgets from a global function.',
+          'Avoid returning widgets from a global function.',
+          'Avoid returning widgets from a global function.',
+          'Avoid returning widgets from a global function.',
         ],
       );
     });